backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Werkzeug==3.0.1
SQLAlchemy==2.0.23
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
authlib==1.2.1
qrcode==7.4.2
geopy==2.4.1
scikit-learn==1.3.2
numpy==1.25.2
bcrypt==4.1.2
marshmallow==3.20.1